supabase/functions/remove-background/requirements.txt

rembg==2.0.50
Pillow==10.0.0
